Use class mixing and combine all stochastic engines in single file
